Title: 토끼와 경주, Time: 6026ms, Memory: 36MB, Status: Time Limit Exceed - CodeTree

+++ 231208/토끼와 경주/rabit-and-race.java
@@ -5,9 +5,9 @@ public class Main {
     
     static class Rabbit implements Comparable<Rabbit>{
         int x, y, pid, d, jump;
-        int score;
+        long score;
 
-        Rabbit(int x, int y, int pid, int d, int jump, int score){
+        Rabbit(int x, int y, int pid, int d, int jump, long score){
             this.x = x;
             this.y = y;
             this.pid = pid;
